data-integration/requirements.txt
gdal==3.0.4
xarray==0.16.2
netcdf4==1.5.5.1
sqlalchemy==1.4.36
requests==2.24.0
psycopg2==2.8.6
geoalchemy2==0.6.3
frictionless==4.0.8
beautifulsoup4==4.9.3
click==7.1.1
pandasdmx==1.5.0
cdsapi==0.5.1
xlrd==1.2.0
openpyxl==3.0.7
clint==0.5.1
shapely==1.7.1 
geopandas>=0.12.1
numpy==1.22.4
rtree==1.0.1
xmltodict==0.13.0
tqdm==4.64.1